fix loading and error display on the map
